+++ src/BubbleSort.h
@@ -4,8 +4,8 @@
 #include <utility>
 
 void bubbleSort(int* arr, int n);
-void bubbleSort(int* arr, int n, int& count_compare);
+void bubbleSort(int* arr, int n, long long& count_compare);
 void bubbleSort_optimized(int* arr, int n);
-void bubbleSort_optimized(int* arr, int n, int& count_compare);
+void bubbleSort_optimized(int* arr, int n, long long& count_compare);
 
 #endif
