+++ src/FlashSort.h
@@ -5,6 +5,6 @@
 #include <algorithm>
 
 void flashSort(int* arr, int n);
-void flashSort(int* arr, int n, int& count_compare);
+void flashSort(int* arr, int n, long long & count_compare);
 
 #endif
