+++ src/radix_sort.h
@@ -5,8 +5,8 @@
 #include <algorithm>
 
 static void __countingSort(int* arr, std::size_t n, int exp);
-static void __countingSort(int* arr, std::size_t n, int exp, unsigned int& cnt);
+static void __countingSort(int* arr, std::size_t n, int exp, long long& cnt);
 void radixSort(int* arr, std::size_t n);
-void radixSort(int* arr, std::size_t n, unsigned int& cnt);
+void radixSort(int* arr, std::size_t n, long long& cnt);
 
 #endif //RADIX_SORT_H
